teachShareAI-project/requirements.txt
flask
flask-cors
python-dotenv
openai
docx
PyPDF2
httpx
